ウィジェットに渡すためのデータ格納と呼び出し
今日のやることのタイトルだけ表示してる
